+++ CHANGELOG.md
@@ -31,6 +31,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Fix the bootstrap date icon visualization
 * Update dark mode and accessibility for poll element
 * Add pin element
+* Add the possibility to change accordion element header background color
+
 
 ## [1.4.9] - 05.02.2025
 * Add new "list/list item" element where it's possible to sort the list with checkbox via drag and drop or keyboard
@@ -43,6 +45,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Fix error in the label and tooltip visualization
 * Fix telephone number form field validation
 * Columns displayed in grid not flex
+
 
 ## [1.4.8] - 11.10.2024
 * export text element with .editor() function where htmlEditorConfig can be customized